Changes to View.cc to make WID and HEI requests with region exports produce images with a *final* size of that requested by WID or HEI. For example RGN=0.1,0.1,0.5,0.5&WID=500 will now produce an image of 500px in width rather than extract the region from an image of size 500px wide. CVT also cleaned up.

+++ src/View.h
@@ -146,27 +146,27 @@ class View{
 
   /// Return the scaling required in case our requested width or height is in between available resolutions
   /* @return scaling factor */
-  double getScale();
+  float getScale();
 
 
   /// Set the left co-ordinate of the viewport
   /** @param x left resolution independent co-ordinate */
-  void setViewLeft( double x );
+  void setViewLeft( float x );
 
 
   /// Set the top co-ordinate of the viewport
   /** @param y top resolution independent co-ordinate */
-  void setViewTop( double y );
+  void setViewTop( float y );
 
 
   /// Set the width co-ordinate of the viewport
   /** @param w width resolution independent co-ordinate */
-  void setViewWidth( double w );
+  void setViewWidth( float w );
 
 
   /// Set the height co-ordinate of the viewport
   /** @param h height resolution independent co-ordinate */
-  void setViewHeight( double h );
+  void setViewHeight( float h );
 
 
   /// Set the source image pixel size
